Move LayerConfiguration and LayerEntry to api package
