+++ jib-core/src/main/java/com/google/cloud/tools/jib/cache/LayerEntriesSelector.java
@@ -17,8 +17,8 @@
 package com.google.cloud.tools.jib.cache;
 
 import com.google.cloud.tools.jib.api.DescriptorDigest;
+import com.google.cloud.tools.jib.api.LayerEntry;
 import com.google.cloud.tools.jib.hash.Digests;
-import com.google.cloud.tools.jib.image.LayerEntry;
 import com.google.cloud.tools.jib.json.JsonTemplate;
 import com.google.common.annotations.VisibleForTesting;
 import com.google.common.collect.ImmutableList;
